format and fix lint/tests
