+++ apps/server/src/services/mediaServer/shared/__tests__/jellyfinEmbyUtils.test.ts
@@ -485,6 +485,7 @@ describe('extractStreamDetails', () => {
           VideoRangeType: undefined,
           VideoRange: undefined,
           ColorSpace: 'bt2020nc',
+          ColorTransfer: undefined,
           BitDepth: 10,
         },
       });
